python_service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
snownlp
jieba
requests
httpx
beautifulsoup4
python-dotenv
apscheduler
bilibili-api-python
cryptography
pydantic-settings
lxml
curl_cffi
DBUtils
# Transformer 情感分析
transformers
torch
sentencepiece
accelerate
pyyaml
scikit-learn
pandas
tqdm
onnxruntime
